+++ tests/test_common/test_utilities/test_command_wrapper.py
@@ -1,7 +1,7 @@
 from unittest import TestCase
 
 from mock import Mock
-from common.wrappers.CommandWrapper import CommandWrapper
+from common.wrappers.command_wrapper import CommandWrapper
 
 
 class TestCommandWrapper(TestCase):
