--- tests/test_vCenterShell.py
@@ -1,8 +1,0 @@
-import unittest
-from mock import Mock
-
-class Test_test_vCenterShell(unittest.TestCase):
-    pass
-
-if __name__ == '__main__':
-    unittest.main()
